requirements.txt
streamlit
pandas
plotly
pdfminer.six
beautifulsoup4
requests